requirements.txt
requests
fake_useragent
bs4
chromadb
langchain
langchain_community
langchain_ollama
langchain_core
langchain_text_splitters
fastembed
langchain_chroma
pdfplumber
easyocr
pillow
numpy
tiktoken
pytesseract
tqdm
